requirements.txt
aiogram>=3.4.1,<4.0.0
motor>=3.4.0,<4.0.0
python-dotenv>=1.0.1,<2.0.0
pymongo>=4.6.0,<5.0.0
